fix(deploy): remove prerequisite only if needed

+++ src/deploy.sh
@@ -87,9 +87,16 @@ fi
 
 # remove prerequesite section if no peer dependencies are found
 if [[ $(cat package.json | jq '.peerDependencies | length') = 0 ]]; then
-  echo "Remove page prerequesite as no peer dependencies were found"
-  jq '.sections |= map(select(.name != "Prerequisite"))' styleguide/styleguide.ext.json > styleguide/styleguide.ext.json.tmp$$
-  mv styleguide/styleguide.ext.json.tmp$$ styleguide/styleguide.ext.json
+  # look for prerequisite file path
+  prerequesiteFile=$(jq '.sections |= map(select(.name == "Prerequisite")) | .sections[0].content' styleguide/styleguide.ext.json)
+  if [[ "$prerequesiteFile" != "null" ]]; then
+    # remove prerequesite if found and $PACKAGE_PEERS is present in file
+    if [[ -f "$prerequesiteFile" ]] && [[ $(grep "\$PACKAGE_PEERS" "$prerequesiteFile") ]]; then
+      echo "Remove page prerequesite as no peer dependencies were found"
+      jq '.sections |= map(select(.name != "Prerequisite"))' styleguide/styleguide.ext.json > styleguide/styleguide.ext.json.tmp$$
+      mv styleguide/styleguide.ext.json.tmp$$ styleguide/styleguide.ext.json
+    fi
+  fi
 fi
 
 if [[ ! -e $PWD/public ]]; then
